Fix bug: data-type

+++ engine/app.py
@@ -38,7 +38,7 @@ def create_app():
         process_name = input_data['Process_name'].iloc[0]
         object_area = input_data['Object_area'].iloc[0]
         hum_count = int(hum_days) // int(input_data['Directive_perfomance'].iloc[0])
-        final_price = int((hum_days * 10) * input_data['Hour_cost'].iloc[0])
+        final_price = int(float((hum_days * 10) * float(input_data['Hour_cost'].iloc[0])))
         hour_cost = input_data['Hour_cost'].iloc[0]
